[wrapper] Use static_cast when casting std::size_t to std::uint32_t. Closes https://github.com/inexorgame/vulkan-renderer/issues/205

+++ src/vulkan-renderer/world/indentation.cpp
@@ -69,6 +69,6 @@ void Indentation::indent_end(std::int8_t steps) noexcept {
 }
 
 std::uint8_t Indentation::uid() const {
-    return 10 * m_start + offset() - (std::pow(m_start, 2) + m_start) / 2;
+    return static_cast<std::uint8_t>(10 * m_start + offset() - (std::pow(m_start, 2) + m_start) / 2);
 }
 } // namespace inexor::vulkan_renderer::world
