+++ src/vulkan-renderer/wrapper/framebuffer.cpp
@@ -15,7 +15,7 @@ Framebuffer::Framebuffer(VkDevice device, VkRenderPass render_pass, const std::v
     spdlog::trace("Creating framebuffer");
 
     auto framebuffer_ci = make_info<VkFramebufferCreateInfo>();
-    framebuffer_ci.attachmentCount = attachments.size();
+    framebuffer_ci.attachmentCount = static_cast<std::uint32_t>(attachments.size());
     framebuffer_ci.pAttachments = attachments.data();
     framebuffer_ci.width = swapchain.get_extent().width;
     framebuffer_ci.height = swapchain.get_extent().height;
